cleanup imports and table vars
