+++ shared/enums/db-tables.enum.ts
@@ -1,4 +1,5 @@
 export enum DatabaseTables {
   USERS = "users",
   ACLS = "acls",
+  RESETS = "resets",
 }
