+++ shared/interfaces/index.ts
@@ -4,4 +4,5 @@ export * from "./base.interface";
 export * from "./invite.interface";
 export * from "./organization.interface";
 export * from "./response.interface";
+export * from "./reset.interface";
 export * from "./user.interface";
