+++ shared/interfaces/reset.interface.ts
@@ -1,0 +1,11 @@
+import { TypesEnum } from "../enums";
+import { ObjectId } from "../types";
+
+export interface PasswordResetInterface {
+  _id?: ObjectId | string;
+  type: TypesEnum.PASSWORD_RESET;
+  id_account: ObjectId | string;
+  email: string;
+  created: Date;
+  expires: Date;
+}
